# ai_agents/memory_systems/hindsight/hindsight-clients/go/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Memory System
  description: HTTP API for Hindsight
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Hindsight HTTP API
  version: 0.4.22
servers:
- url: /
paths:
...
  /v1/default/banks/{bank_id}/memories/list:
    get:
      description: "List memory units with pagination and optional full-text search.\
        \ Supports filtering by type. Results are sorted by most recent first (mentioned_at\
        \ DESC, then created_at DESC)."
      operationId: list_memories
      parameters:
      - explode: false
        in: path
        name: bank_id
        required: true
        schema:
          title: Bank Id
          type: string
        style: simple
      - explode: true
        in: query
        name: type
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: q
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
        style: form
      - explode: false
        in: header
        name: authorization
        required: false
        schema:
          nullable: true
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMemoryUnitsResponse'
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List memory units
      tags:
      - Memory
...
components:
  schemas:
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
          url: url
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
          url: url
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
      title: HTTPValidationError
...
    ListMemoryUnitsResponse:
      description: Response model for list memory units endpoint.
      example:
        items:
        - context: Work conversation
          date: 2024-01-15T10:30:00Z
          entities: "Alice (PERSON), Google (ORGANIZATION)"
          id: 550e8400-e29b-41d4-a716-446655440000
          text: Alice works at Google on the AI team
          type: world
        limit: 100
        offset: 0
        total: 150
      properties:
        items:
          items:
            additionalProperties: {}
          type: array
        total:
          title: Total
          type: integer
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
      required:
      - items
      - limit
      - offset
      - total
      title: ListMemoryUnitsResponse
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        input: ""
        ctx: "{}"
        type: type
        url: url
      properties:
        loc:
          items:
            $ref: '#/components/schemas/ValidationError_loc_inner'
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
        input: {}
        ctx:
          title: Context
          type: object
        url:
          title: URL
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer